+++ python/lsst/ts/IntegrationTests/configs/maintel_housekeeping_configs.py
@@ -38,7 +38,7 @@ yaml_string = yaml.safe_load(
     """
 )
 
-registry["cccamera_housekeeping"] = yaml.safe_dump(
+registry["cccamera_set_filter"] = yaml.safe_dump(
     yaml_string,
     explicit_start=True,
     canonical=True,
@@ -52,7 +52,40 @@ yaml_string = yaml.safe_load(
     """
 )
 
-registry["mtmount_housekeeping"] = yaml.safe_dump(
+registry["mtmount_home_both_axes"] = yaml.safe_dump(
+    yaml_string,
+    explicit_start=True,
+    canonical=True,
+)
+
+# MTPtg park
+yaml_string = yaml.safe_load(
+    """
+    component: "MTPtg"
+    cmd: azElTarget
+    parameters:
+        targetName: Park position
+        azDegs: 0
+        elDegs: 80
+        rotPA: 0
+    """
+)
+
+registry["mtptg_park"] = yaml.safe_dump(
+    yaml_string,
+    explicit_start=True,
+    canonical=True,
+)
+
+# MTPtg stop tracking
+yaml_string = yaml.safe_load(
+    """
+    component: "MTPtg"
+    cmd: stopTracking
+    """
+)
+
+registry["mtptg_stop_tracking"] = yaml.safe_dump(
     yaml_string,
     explicit_start=True,
     canonical=True,