+++ python/lsst/ts/IntegrationTests/maintel_housekeeping.py
@@ -38,10 +38,14 @@ class MainTelHousekeeping(BaseScript):
 
     index: int = 1
     configs: tuple = (
-        registry["cccamera_housekeeping"],
-        registry["mtmount_housekeeping"],
+        registry["cccamera_set_filter"],
+        registry["mtmount_home_both_axes"],
+        registry["mtptg_park"],
+        registry["mtptg_stop_tracking"],
     )
     scripts: list = [
+        ("run_command.py", BaseScript.is_standard),
+        ("run_command.py", BaseScript.is_standard),
         ("run_command.py", BaseScript.is_standard),
         ("run_command.py", BaseScript.is_standard),
     ]