+++ tests/test_maintel_housekeeping.py
@@ -55,7 +55,7 @@ class MainTelHousekeepingTestCase(unittest.IsolatedAsyncioTestCase):
         # Assert script was added to ScriptQueue.
         self.assertEqual(len(self.controller.queue_list), num_scripts)
         # Assert scripts passed.
-        self.assertEqual(script_class.script_states, [8, 8])
+        self.assertEqual(script_class.script_states, [8, 8, 8, 8])
 
     async def asyncTearDown(self) -> None:
         await self.controller.close()