+++ doc/version-history.rst
@@ -10,6 +10,12 @@ Version History
 .. No new work should be required in order to complete this section.
 .. Below is an example of a version history format.
 
+v0.16.0
+-------
+* Added Watcher to the ObsSys State transition tests.
+* Added MTPtg park and stop tracking tests to MainTel Housekeeping.
+* Consolidated the ObsSys 1 & 2 tests into just ObsSys.
+
 v0.15.0
 -------
 * Added the ESS:106 CSC.